Remove reference to autorest from arm library
