+++ packages/typespec-azure-resource-manager/lib/common-types/private-links.tsp
@@ -2,7 +2,6 @@ import "@typespec/openapi";
 import "@typespec/http";
 import "@typespec/rest";
 import "@typespec/versioning";
-import "@azure-tools/typespec-autorest";
 import "@azure-tools/typespec-azure-core";
 
 using TypeSpec.Http;
