+++ packages/typespec-azure-resource-manager/package.json
@@ -56,7 +56,6 @@
     "pluralize": "^8.0.0"
   },
   "peerDependencies": {
-    "@azure-tools/typespec-autorest": "workspace:~",
     "@azure-tools/typespec-azure-core": "workspace:~",
     "@typespec/compiler": "workspace:~",
     "@typespec/http": "workspace:~",
@@ -65,7 +64,6 @@
     "@typespec/versioning": "workspace:~"
   },
   "devDependencies": {
-    "@azure-tools/typespec-autorest": "workspace:~",
     "@azure-tools/typespec-azure-core": "workspace:~",
     "@types/node": "~18.11.19",
     "@types/pluralize": "^0.0.33",
