+++ packages/typespec-azure-resource-manager/src/index.ts
@@ -2,7 +2,16 @@ import { definePackageFlags } from "@typespec/compiler";
 
 export const namespace = "Azure.ResourceManager";
 
-export * from "./common-types.js";
+export {
+  $armCommonTypesVersion,
+  getArmCommonTypeOpenAPIRef,
+  getArmCommonTypesVersion,
+  getArmCommonTypesVersions,
+  isArmCommonType,
+  type ArmCommonTypeVersions,
+  type ArmCommonTypesResolutionOptions,
+} from "./common-types.js";
+
 export * from "./namespace.js";
 export * from "./operations.js";
 export * from "./resource.js";
