+++ packages/typespec-azure-resource-manager/test/resource.test.ts
@@ -1,10 +1,10 @@
-import { Diagnostic, Model } from "@typespec/compiler";
+import { Model } from "@typespec/compiler";
 import { expectDiagnosticEmpty, expectDiagnostics } from "@typespec/compiler/testing";
-import { deepEqual, ok, strictEqual } from "assert";
+import { ok, strictEqual } from "assert";
 import { describe, it } from "vitest";
 import { ArmLifecycleOperationKind } from "../src/operations.js";
 import { ArmResourceDetails, getArmResources } from "../src/resource.js";
-import { checkFor, getOpenApiAndDiagnostics, openApiFor } from "./test-host.js";
+import { checkFor } from "./test-host.js";
 
 function assertLifecycleOperation(
   resource: ArmResourceDetails,
@@ -627,453 +627,6 @@ describe("typespec-azure-resource-manager: ARM resource model", () => {
           code: "@azure-tools/typespec-azure-resource-manager/arm-resource-missing-name-segment-decorator",
         });
       });
-    });
-
-    describe("emits correct openApi:", () => {
-      it("emits correct paths for tenant resources", async () => {
-        const openApi = await openApiFor(`
-        @armProviderNamespace
-        @useDependency(Azure.ResourceManager.Versions.v1_0_Preview_1)
-        namespace Microsoft.Contoso;
-
-        @doc("Widget resource")
-        model Widget is ProxyResource<WidgetProperties> {
-          @doc("The name of the widget")
-          @key("widgetName")
-          @segment("widgets")
-          @path
-          @visibility("read")
-          name: string;
-        }
-
-      @doc("The properties of a widget")
-      model WidgetProperties {
-        #suppress "deprecated" "for testing"
-        @doc("The spin of the widget")
-        @knownValues(SpinValues)
-        spin?: string;
-      }
-      
-      enum SpinValues {
-        up,
-        down,
-      }
-      
-      interface Widgets extends TenantResourceOperations<Widget, WidgetProperties> {
-        @autoRoute
-        @doc("Flip to the opposite of the current spin")
-        @action
-        flipSpin(...TenantInstanceParameters<Widget>): ArmResponse<Widget> | ErrorResponse;
-      }
-
-      @doc("Flange resource")
-      @parentResource(Widget)
-      model Flange is ProxyResource<FlangeProperties> {
-        @doc("The name of the flange")
-        @key("flangeName")
-        @segment("flanges")
-        @path
-        @visibility("read")
-        name: string;
-      }
-      
-      @doc("The properties of a Flange")
-      model FlangeProperties {
-        @doc("The description of the flange")
-        description: string;
-      
-        @doc("The weight in ounces of the flange")
-        weight: safeint;
-      }
-      
-      #suppress "deprecated" "test"
-      interface Flanges
-        extends TenantResourceCreate<Flange>,
-          TenantResourceDelete<Flange>,
-          TenantResourceRead<Flange>,
-          TenantResourceListByParent<Flange> {
-        @autoRoute
-        @doc("Add a certain amount to the weight")
-        @action
-        increaseWeight(
-          ...TenantInstanceParameters<Flange>,
-          increment: safeint
-        ): ArmNoContentResponse<"Weight added successfully"> | ErrorResponse;
-      }
-      `);
-        ok(openApi.paths["/providers/Microsoft.Contoso/widgets"].get);
-        ok(openApi.paths["/providers/Microsoft.Contoso/widgets/{widgetName}"].get);
-        ok(openApi.paths["/providers/Microsoft.Contoso/widgets/{widgetName}"].put);
-        ok(openApi.paths["/providers/Microsoft.Contoso/widgets/{widgetName}"].patch);
-        ok(openApi.paths["/providers/Microsoft.Contoso/widgets/{widgetName}"].delete);
-        ok(openApi.paths["/providers/Microsoft.Contoso/widgets/{widgetName}/flanges"].get);
-        ok(
-          openApi.paths["/providers/Microsoft.Contoso/widgets/{widgetName}/flanges/{flangeName}"]
-            .get
-        );
-        ok(
-          openApi.paths["/providers/Microsoft.Contoso/widgets/{widgetName}/flanges/{flangeName}"]
-            .put
-        );
-        ok(
-          openApi.paths["/providers/Microsoft.Contoso/widgets/{widgetName}/flanges/{flangeName}"]
-            .delete
-        );
-        ok(
-          openApi.paths[
-            "/providers/Microsoft.Contoso/widgets/{widgetName}/flanges/{flangeName}/increaseWeight"
-          ].post
-        );
-      });
-
-      it("emits correct paths for checkLocalName endpoints", async () => {
-        const openApi = await openApiFor(`
-          @armProviderNamespace
-          @useDependency(Azure.ResourceManager.Versions.v1_0_Preview_1)
-          namespace Microsoft.Contoso;
-
-          @doc("Widget resource")
-          model Widget is ProxyResource<WidgetProperties> {
-            @doc("The name of the widget")
-            @key("widgetName")
-            @segment("widgets")
-            @path
-            @visibility("read")
-            name: string;
-          }
-
-          @doc("The properties of a widget")
-          model WidgetProperties {
-            #suppress "deprecated" "for testing"
-            @doc("The spin of the widget")
-            @knownValues(SpinValues)
-            spin?: string;
-          }
-      
-          enum SpinValues {
-            up,
-            down,
-          }
-
-          interface Widgets extends Operations {
-            checkName is checkGlobalNameAvailability;
-            checkLocalName is checkLocalNameAvailability;
-          }
-      `);
-        ok(
-          openApi.paths[
-            "/subscriptions/{subscriptionId}/providers/Microsoft.Contoso/checkNameAvailability"
-          ].post
-        );
-        ok(
-          openApi.paths[
-            "/subscriptions/{subscriptionId}/providers/Microsoft.Contoso/locations/{location}/checkNameAvailability"
-          ].post
-        );
-      });
-
-      describe("emits refs for common-types at selected version or earlier versions", () => {
-        function assertRef(item: any, refPath: string) {
-          ok("$ref" in item);
-          strictEqual(item["$ref"], refPath);
-        }
-
-        function assertCommonDef(
-          openApi: any,
-          path: string,
-          defName: string,
-          version: string,
-          file: string
-        ) {
-          assertRef(
-            openApi.paths[path].get.responses["200"].schema,
-            `../../common-types/resource-management/${version}/${file}#/definitions/${defName}`
-          );
-        }
-
-        function assertCommonParam(
-          openApi: any,
-          path: string,
-          paramName: string,
-          version: string,
-          file: string
-        ) {
-          assertRef(
-            openApi.paths[path].get.parameters[0],
-            `../../common-types/resource-management/${version}/${file}#/parameters/${paramName}`
-          );
-        }
-
-        function assertNotCommonDef(openApi: any, path: string, defName: string) {
-          assertRef(openApi.paths[path].get.responses["200"].schema, `#/definitions/${defName}`);
-        }
-
-        function assertNotCommonParam(openApi: any, path: string, paramName: string) {
-          assertRef(openApi.paths[path].get.parameters[0], `#/parameters/${paramName}`);
-        }
-
-        function testCommonTypesVersion(
-          version: string | undefined,
-          assertFunc: (openApi: any) => void,
-          expectedDiagnosticsPerVersion: Partial<Diagnostic>[]
-        ) {
-          async function runTest(useVersionEnum: boolean): Promise<void> {
-            const versions = useVersionEnum ? ["v1", "v2"] : undefined;
-            const decorators = `@useDependency(Azure.ResourceManager.Versions.v1_0_Preview_1) ${
-              version
-                ? `@armCommonTypesVersion(Azure.ResourceManager.CommonTypes.Versions.${version})`
-                : ""
-            }`;
-            const nsDecorators = useVersionEnum
-              ? `@versioned(Microsoft.Contoso.Versions)`
-              : decorators;
-            const versionEnum = useVersionEnum
-              ? `enum Versions { ${decorators} v1, ${decorators} v2 }`
-              : "";
-
-            const [openApi, diagnostics] = await getOpenApiAndDiagnostics(
-              `
-              @armProviderNamespace
-              @service({ title: "Contoso" })
-              ${nsDecorators}
-              namespace Microsoft.Contoso;
-
-              ${versionEnum}
-
-              @Azure.ResourceManager.Private.armCommonDefinition("Foo", Azure.ResourceManager.CommonTypes.Versions.v3, "foo.json")
-              model Foo {prop: string}
-
-              model FooParam {
-                @path
-                @Azure.ResourceManager.Private.armCommonParameter("FooParam", Azure.ResourceManager.CommonTypes.Versions.v3, "foo.json")
-                foo: string;
-              }
-
-              @Azure.ResourceManager.Private.armCommonDefinition("Bar", #{ version: Azure.ResourceManager.CommonTypes.Versions.v4, isDefault: true }, "bar.json")
-              @Azure.ResourceManager.Private.armCommonDefinition("Bar", Azure.ResourceManager.CommonTypes.Versions.v5, "bar-v5.json")
-              model Bar {prop: string}
-
-              model BarParam {
-                @path
-                @Azure.ResourceManager.Private.armCommonParameter("BarParam", #{ version: Azure.ResourceManager.CommonTypes.Versions.v4, isDefault: true }, "bar.json")
-                @Azure.ResourceManager.Private.armCommonParameter("BarParam", Azure.ResourceManager.CommonTypes.Versions.v5, "bar-v5.json")
-                bar: string;
-              }
-
-              @Azure.ResourceManager.Private.armCommonDefinition("Baz", Azure.ResourceManager.CommonTypes.Versions.v5, "baz.json")
-              model Baz {prop: string}
-
-              model BazParam {
-                @path
-                @Azure.ResourceManager.Private.armCommonParameter("BazParam", Azure.ResourceManager.CommonTypes.Versions.v5, "baz.json")
-                baz: string;
-              }
-
-              @route("/foo") op getFoo(...FooParam): Foo;
-              @route("/bar") op getBar(...BarParam): Bar;
-              @route("/baz") op getBaz(...BazParam): Baz;
-              `,
-              {},
-              versions
-            );
-
-            // The same diagnostics will be raised each version the spec is
-            // generated for so dynamically generate the expected diagnostic
-            // array
-            expectDiagnostics(
-              diagnostics,
-              Array(versions ? versions.length : 1)
-                .fill(expectedDiagnosticsPerVersion)
-                .flat(1)
-            );
-
-            assertFunc(useVersionEnum ? openApi.v2 : openApi);
-          }
-
-          it("with version enum", async () => {
-            await runTest(true);
-          });
-          it("with version on namespace", async () => {
-            await runTest(false);
-          });
-        }
-
-        describe("v3", () => {
-          testCommonTypesVersion(
-            "v3",
-            (openApi) => {
-              assertCommonDef(openApi, "/foo/{foo}", "Foo", "v3", "foo.json");
-              assertCommonParam(openApi, "/foo/{foo}", "FooParam", "v3", "foo.json");
-
-              assertNotCommonDef(openApi, "/bar/{bar}", "Bar");
-              assertNotCommonParam(openApi, "/bar/{bar}", "BarParam");
-              assertNotCommonDef(openApi, "/baz/{baz}", "Baz");
-              assertNotCommonParam(openApi, "/baz/{baz}", "BazParam");
-            },
-            [
-              {
-                code: "@azure-tools/typespec-azure-resource-manager/arm-common-types-incompatible-version",
-                message:
-                  "No ARM common-types version for this type satisfies the expected version v3.  This type only supports the following version(s): v5, v4",
-              },
-              {
-                code: "@azure-tools/typespec-azure-resource-manager/arm-common-types-incompatible-version",
-                message:
-                  "No ARM common-types version for this type satisfies the expected version v3.  This type only supports the following version(s): v5, v4",
-              },
-              {
-                code: "@azure-tools/typespec-azure-resource-manager/arm-common-types-incompatible-version",
-                message:
-                  "No ARM common-types version for this type satisfies the expected version v3.  This type only supports the following version(s): v5",
-              },
-              {
-                code: "@azure-tools/typespec-azure-resource-manager/arm-common-types-incompatible-version",
-                message:
-                  "No ARM common-types version for this type satisfies the expected version v3.  This type only supports the following version(s): v5",
-              },
-            ]
-          );
-        });
-
-        describe("v4", () => {
-          testCommonTypesVersion(
-            "v4",
-            (openApi) => {
-              assertCommonDef(openApi, "/foo/{foo}", "Foo", "v3", "foo.json");
-              assertCommonParam(openApi, "/foo/{foo}", "FooParam", "v3", "foo.json");
-              assertCommonDef(openApi, "/bar/{bar}", "Bar", "v4", "bar.json");
-              assertCommonParam(openApi, "/bar/{bar}", "BarParam", "v4", "bar.json");
-
-              assertNotCommonDef(openApi, "/baz/{baz}", "Baz");
-              assertNotCommonParam(openApi, "/baz/{baz}", "BazParam");
-            },
-            [
-              {
-                code: "@azure-tools/typespec-azure-resource-manager/arm-common-types-incompatible-version",
-                message:
-                  "No ARM common-types version for this type satisfies the expected version v4.  This type only supports the following version(s): v5",
-              },
-              {
-                code: "@azure-tools/typespec-azure-resource-manager/arm-common-types-incompatible-version",
-                message:
-                  "No ARM common-types version for this type satisfies the expected version v4.  This type only supports the following version(s): v5",
-              },
-            ]
-          );
-        });
-
-        describe("v5", () => {
-          testCommonTypesVersion(
-            "v5",
-            (openApi) => {
-              assertCommonDef(openApi, "/foo/{foo}", "Foo", "v3", "foo.json");
-              assertCommonParam(openApi, "/foo/{foo}", "FooParam", "v3", "foo.json");
-              assertCommonDef(openApi, "/bar/{bar}", "Bar", "v5", "bar-v5.json");
-              assertCommonParam(openApi, "/bar/{bar}", "BarParam", "v5", "bar-v5.json");
-              assertCommonDef(openApi, "/baz/{baz}", "Baz", "v5", "baz.json");
-              assertCommonParam(openApi, "/baz/{baz}", "BazParam", "v5", "baz.json");
-            },
-            []
-          );
-        });
-
-        describe("unspecified version (use defaults)", () => {
-          testCommonTypesVersion(
-            undefined,
-            (openApi) => {
-              assertCommonDef(openApi, "/foo/{foo}", "Foo", "v3", "foo.json");
-              assertCommonParam(openApi, "/foo/{foo}", "FooParam", "v3", "foo.json");
-              assertCommonDef(openApi, "/bar/{bar}", "Bar", "v4", "bar.json");
-              assertCommonParam(openApi, "/bar/{bar}", "BarParam", "v4", "bar.json");
-
-              assertNotCommonDef(openApi, "/baz/{baz}", "Baz");
-              assertNotCommonParam(openApi, "/baz/{baz}", "BazParam");
-            },
-            [
-              // Only the v5 type gets complained about because the v4 type has isDefault: true
-              {
-                code: "@azure-tools/typespec-azure-resource-manager/arm-common-types-incompatible-version",
-                message:
-                  "No ARM common-types version for this type satisfies the expected version v3.  This type only supports the following version(s): v5",
-              },
-              {
-                code: "@azure-tools/typespec-azure-resource-manager/arm-common-types-incompatible-version",
-                message:
-                  "No ARM common-types version for this type satisfies the expected version v3.  This type only supports the following version(s): v5",
-              },
-            ]
-          );
-        });
-      });
-    });
-  });
-
-  it("emits correct paths for ArmResourceHead operation", async () => {
-    const openApi = await openApiFor(`
-        @armProviderNamespace
-        @armCommonTypesVersion(Azure.ResourceManager.CommonTypes.Versions.v5)
-        @useDependency(Azure.ResourceManager.Versions.v1_0_Preview_1)
-        namespace Microsoft.Contoso;
-        @doc("Widget resource")
-        model Widget is ProxyResource<WidgetProperties> {
-          @doc("The name of the widget")
-          @key("widgetName")
-          @segment("widgets")
-          @path
-          @visibility("read")
-          name: string;
-        }
-        @doc("The properties of a widget")
-        model WidgetProperties {
-          #suppress "deprecated" "for testing"
-          @doc("The spin of the widget")
-          @knownValues(SpinValues)
-          spin?: string;
-        }
-        enum SpinValues {
-          up,
-          down,
-        }
-        interface Widgets extends Operations {
-          checkExist is ArmResourceCheckExistence<Widget>;
-        }
-    `);
-    const headOperation =
-      openApi.paths[
-        "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Contoso/widgets/{widgetName}"
-      ].head;
-    ok(headOperation);
-    deepEqual(headOperation, {
-      operationId: "Widgets_CheckExist",
-      description: "Check for the existence of a Widget",
-      parameters: [
-        {
-          $ref: "../../common-types/resource-management/v5/types.json#/parameters/ApiVersionParameter",
-        },
-        {
-          $ref: "../../common-types/resource-management/v5/types.json#/parameters/SubscriptionIdParameter",
-        },
-        {
-          $ref: "../../common-types/resource-management/v5/types.json#/parameters/ResourceGroupNameParameter",
-        },
-        {
-          name: "widgetName",
-          in: "path",
-          description: "The name of the widget",
-          required: true,
-          type: "string",
-        },
-      ],
-      responses: {
-        "204": { description: "The Azure resource exists" },
-        "404": { description: "The Azure resource is not found" },
-        default: {
-          description: "An unexpected error response.",
-          schema: {
-            $ref: "../../common-types/resource-management/v5/types.json#/definitions/ErrorResponse",
-          },
-        },
-      },
     });
   });
