+++ packages/typespec-azure-resource-manager/test/test-host.ts
@@ -1,18 +1,10 @@
-import { AutorestEmitterOptions } from "@azure-tools/typespec-autorest";
-import { AutorestTestLibrary } from "@azure-tools/typespec-autorest/testing";
 import { AzureCoreTestLibrary } from "@azure-tools/typespec-azure-core/testing";
 import { Diagnostic, Program } from "@typespec/compiler";
-import {
-  createTestHost,
-  createTestWrapper,
-  expectDiagnosticEmpty,
-  resolveVirtualPath,
-} from "@typespec/compiler/testing";
+import { createTestHost, createTestWrapper } from "@typespec/compiler/testing";
 import { HttpTestLibrary } from "@typespec/http/testing";
 import { OpenAPITestLibrary } from "@typespec/openapi/testing";
 import { RestTestLibrary } from "@typespec/rest/testing";
 import { VersioningTestLibrary } from "@typespec/versioning/testing";
-import { OpenAPI2Document } from "../../typespec-autorest/src/openapi2-document.js";
 import { $lib } from "../src/lib.js";
 import { AzureResourceManagerTestLibrary } from "../src/testing/index.js";
 
@@ -22,7 +14,6 @@ export async function createAzureResourceManagerTestHost() {
       HttpTestLibrary,
       RestTestLibrary,
       OpenAPITestLibrary,
-      AutorestTestLibrary,
       AzureCoreTestLibrary,
       VersioningTestLibrary,
       AzureResourceManagerTestLibrary,
@@ -35,56 +26,6 @@ export async function createAzureResourceManagerTestRunner() {
   return createTestWrapper(host, {
     autoUsings: [`Azure.ResourceManager`, `TypeSpec.Http`, `TypeSpec.Rest`, `TypeSpec.Versioning`],
   });
-}
-
-export async function getOpenApiAndDiagnostics(
-  code: string,
-  options: AutorestEmitterOptions = {},
-  versions?: string[]
-): Promise<[any, readonly Diagnostic[]]> {
-  const runner = await createAzureResourceManagerTestRunner();
-  const diagnostics = await runner.diagnose(code, {
-    noEmit: false,
-    emitters: {
-      [AutorestTestLibrary.name]: {
-        ...options,
-        "emitter-output-dir": resolveVirtualPath("tsp-output"),
-      },
-    },
-  });
-
-  if (versions) {
-    const output: any = {};
-    for (const version of versions) {
-      output[version] = JSON.parse(
-        runner.fs.get(resolveVirtualPath("tsp-output", version, "openapi.json"))!
-      );
-    }
-    return [output, diagnostics];
-  }
-
-  const outPath = resolveVirtualPath("tsp-output", "openapi.json");
-  return [JSON.parse(runner.fs.get(outPath)!), diagnostics];
-}
-
-export async function openApiFor(
-  code: string,
-  options: AutorestEmitterOptions = {},
-  versions?: string[]
-): Promise<any> {
-  const [openApi, diagnostics] = await getOpenApiAndDiagnostics(code, options, versions);
-  expectDiagnosticEmpty(diagnostics);
-  return openApi;
-}
-
-export async function openApiForVersions<T extends string>(
-  code: string,
-  versions: T[],
-  options: AutorestEmitterOptions = {}
-): Promise<Record<T, OpenAPI2Document>> {
-  const [openApi, diagnostics] = await getOpenApiAndDiagnostics(code, options, versions);
-  expectDiagnosticEmpty(diagnostics);
-  return openApi;
 }
 
 export async function checkFor(
